ml-service/requirements.txt
numpy>=1.23.5
pandas>=1.5.0
scikit-learn>=0.24.2
xgboost>=1.4.2
tensorflow>=2.6.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
requests>=2.26.0
beautifulsoup4>=4.9.3
fake-useragent>=0.1.11
tenacity>=8.0.1
python-multipart>=0.0.5
pydantic>=1.8.2
plotly>=5.3.1
jinja2>=3.0.1
pytest-cov>=2.12.1
pytest-asyncio>=0.15.1
httpx>=0.23.0
redis>=4.0.0
prometheus-client>=0.12.0